# docs/contracts/social-media-v2-openapi.json
{
  "components": {
    "schemas": {
...
      "BrandFamilyProjection": {
        "properties": {
          "brand_ids": {
            "items": {
              "type": "string"
            },
            "title": "Brand Ids",
            "type": "array"
          },
          "root_brand_id": {
            "title": "Root Brand Id",
            "type": "string"
          }
        },
        "required": [
          "root_brand_id",
          "brand_ids"
        ],
        "title": "BrandFamilyProjection",
        "type": "object"
      },
...
      "BrandScope": {
        "properties": {
          "requested_brand_id": {
            "title": "Requested Brand Id",
            "type": "string"
          },
          "resolved_brand_ids": {
            "items": {
              "type": "string"
            },
            "title": "Resolved Brand Ids",
            "type": "array"
          },
          "rollup": {
            "title": "Rollup",
            "type": "boolean"
          }
        },
        "required": [
          "requested_brand_id",
          "rollup",
          "resolved_brand_ids"
        ],
        "title": "BrandScope",
        "type": "object"
      },
      "BrandWorkspace": {
        "properties": {
          "brands": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceBrand"
            },
            "title": "Brands",
            "type": "array"
          },
          "default_brand_id": {
            "title": "Default Brand Id",
            "type": "string"
          },
          "families": {
            "items": {
              "$ref": "#/components/schemas/BrandFamilyProjection"
            },
            "title": "Families",
            "type": "array"
          },
          "scope": {
            "$ref": "#/components/schemas/BrandScope"
          }
        },
        "required": [
          "default_brand_id",
          "brands",
          "families",
          "scope"
        ],
        "title": "BrandWorkspace",
        "type": "object"
      },
      "CapabilityId": {
        "enum": [
          "profile",
          "content",
          "comments",
          "audience"
        ],
        "title": "CapabilityId",
        "type": "string"
      },
      "CapabilityPlatform": {
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/CapabilityRecord"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "linked_account_count": {
            "title": "Linked Account Count",
            "type": "integer"
          },
          "navigation_available": {
            "title": "Navigation Available",
            "type": "boolean"
          },
          "platform": {
            "$ref": "#/components/schemas/PlatformId"
          }
        },
        "required": [
          "platform",
          "capabilities",
          "linked_account_count",
          "navigation_available"
        ],
        "title": "CapabilityPlatform",
        "type": "object"
      },
      "CapabilityRecord": {
        "properties": {
          "capability": {
            "$ref": "#/components/schemas/CapabilityId"
          },
          "platform": {
            "$ref": "#/components/schemas/PlatformId"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CapabilityStatus"
          }
        },
        "required": [
          "platform",
          "capability",
          "status",
          "reason"
        ],
        "title": "CapabilityRecord",
        "type": "object"
      },
      "CapabilityStatus": {
        "enum": [
          "unsupported",
          "not_approved",
          "not_configured",
          "blocked_configuration",
          "manual_activation_required",
          "partial",
          "available"
        ],
        "title": "CapabilityStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlatformId": {
        "enum": [
          "facebook",
          "instagram",
          "tiktok"
        ],
        "title": "PlatformId",
        "type": "string"
      },
...
      "RuntimeCapabilities": {
        "properties": {
          "automated_schedule_available": {
            "title": "Automated Schedule Available",
            "type": "boolean"
          },
          "mode": {
            "$ref": "#/components/schemas/RuntimeMode"
          },
          "writes_enabled": {
            "title": "Writes Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "mode",
          "writes_enabled",
          "automated_schedule_available"
        ],
        "title": "RuntimeCapabilities",
        "type": "object"
      },
      "RuntimeMode": {
        "enum": [
          "development",
          "dormant",
          "cutover_read_only",
          "cutover_credential_migration",
          "cutover_canary",
          "cutover_control_plane_drain",
          "cutover_activation",
          "active"
        ],
        "title": "RuntimeMode",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkspaceBrand": {
        "properties": {
          "access_mode": {
            "anyOf": [
              {
                "enum": [
                  "read",
                  "write"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Mode"
          },
          "brand_id": {
            "title": "Brand Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "parent_brand_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Brand Id"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "visibility": {
            "enum": [
              "active",
              "hidden_parent"
            ],
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "brand_id",
          "name",
          "parent_brand_id",
          "visibility",
          "access_mode",
          "role"
        ],
        "title": "WorkspaceBrand",
        "type": "object"
      },
      "WorkspaceCapabilitiesResponse": {
        "properties": {
          "permissions": {
            "$ref": "#/components/schemas/WorkspacePermissions"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/CapabilityPlatform"
            },
            "title": "Platforms",
            "type": "array"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeCapabilities"
          },
          "scope": {
            "$ref": "#/components/schemas/BrandScope"
          }
        },
        "required": [
          "scope",
          "platforms",
          "permissions",
          "runtime"
        ],
        "title": "WorkspaceCapabilitiesResponse",
        "type": "object"
      },
      "WorkspacePermissions": {
        "properties": {
          "internal_audit_visible": {
            "title": "Internal Audit Visible",
            "type": "boolean"
          },
          "meta_connection_manage": {
            "title": "Meta Connection Manage",
            "type": "boolean"
          },
          "operation_mutation_available": {
            "title": "Operation Mutation Available",
            "type": "boolean"
          },
          "rollup_available": {
            "title": "Rollup Available",
            "type": "boolean"
          },
          "settings_visible": {
            "title": "Settings Visible",
            "type": "boolean"
          },
          "tiktok_connection_manage": {
            "title": "Tiktok Connection Manage",
            "type": "boolean"
          }
        },
        "required": [
          "settings_visible",
          "internal_audit_visible",
          "rollup_available",
          "operation_mutation_available",
          "tiktok_connection_manage",
          "meta_connection_manage"
        ],
        "title": "WorkspacePermissions",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Social Media V2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workspace/brands": {
      "get": {
        "operationId": "workspace_brands_api_workspace_brands_get",
        "parameters": [
          {
            "in": "query",
            "name": "selected_brand_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Selected Brand Id"
            }
          },
          {
            "in": "query",
            "name": "rollup",
            "required": false,
            "schema": {
              "default": false,
              "title": "Rollup",
              "type": "boolean"
            }
          },
          {
            "in": "cookie",
            "name": "social_media_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Social Media Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandWorkspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace Brands"
      }
    },
    "/api/workspace/capabilities": {
      "get": {
        "operationId": "workspace_capabilities_api_workspace_capabilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "selected_brand_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Selected Brand Id"
            }
          },
          {
            "in": "query",
            "name": "rollup",
            "required": false,
            "schema": {
              "default": false,
              "title": "Rollup",
              "type": "boolean"
            }
          },
          {
            "in": "cookie",
            "name": "social_media_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Social Media Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceCapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace Capabilities"
      }
    }
  }
}